Extract TUI plugin catalog rendering
